update call leads flow
No longer emiting stopCallingLeads action when getting a "hangup bc call is in progress" reply.

Instead we are just neglecting to call the next lead, but leaving the callingLeads flag in the redux state as true. This keeps the buttons in the header from briefly switching to "resume calling" in the period between receiving the first hangup and the "call answered" reply.